style(Formatting): Formatted and cleaned comments
Also added nother tip
